// sprint-02/FE/src/app/model/books.ts
export interface Books {
  id?: number;
  name?: string;
  image?: string;
  supplier?: string;
  author?: string;
  translator?: string;
  publisher?: string;
  publishYear?: string;
  weight?: number;
  demension?: string;
  pages?: number;
  price?: number;
  cover?: string;
  stock?: number;
  details?: string;
}